// umc_week3_misson23/src/types/movie.ts
export type Movie = {
    adult: boolean;
    backdrop_path: string;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
  };
  
  export type MovieResponse = {
    page: number;
    results: Movie[]; // 실제로 들어오는거는 여러개의 영화 데이터니 Movie의 배열로 표현
    total_pages: number;
    total_results: number;
  };
  
export type MovieDetail = {

    adult: boolean;
    backdrop_path: string;
    belongs_to_collection: null | string;
    budget: number;
    genres: {id:number; name:string}[];
    homepage: string;
    id: number;
    imdb_id: string;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string;
    production_companies:{id:number; logo_path:string; name:string; origin_country:string}[];
    production_countries:{iso_3166_1:string; name:string}[];
    release_date: string;
    revenue: number;
    runtime: number;
    spoken_languages:{english_name:string; iso_639_1:string; name:string}[];
    status: string;
    tagline: string;
    title: string;
    video:boolean
    vote_average: number;
    vote_count: number;
   

}

export type MovieCredits = {
    id: number;
    cast: {
        adult: boolean;gender: number; id: number; known_for_department: string; name: string; original_name: string; popularity: number; profile_path: string; cast_id: number; character: string; credit_id: string; order: number}[];
    crew:{adult: boolean;gender: number; id: number; known_for_department: string; name: string; original_name: string; popularity: number; profile_path: string; credit_id: string; department: string; job: string}[];

    }



 